tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-delay_two_values.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-delay value with two list entries</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-delay-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="400ms, 0.8s is a valid animation-delay value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes darken-left {
         from {
            border-left-color: blue;
         }
         to {
            border-left-color: black;
         }
      }
      @keyframes darken-right {
         from {
            border-right-color: blue;
         }
         to {
            border-right-color: black;
         }
      }
      div {
         width: 0px;
         height: 100px;
         border-left: 50px solid blue;
         border-right: 50px solid blue;
      }
      .test {
         animation-duration: 1s;
         animation-delay: 400ms, 0.8s;
         animation-fill-mode: forwards;
         animation-timing-function: step-start;
         animation-name: darken-left, darken-right;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var passed = false;
         function aboutEqual(observed, actual) {
            var maxDeviation = 100;
            return (Math.abs(observed - actual) < maxDeviation);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function(evt) {
            var delaySeen = Date.now() - delayStart;
            if (evt.animationName == 'darken-left') {
               passed = aboutEqual(delaySeen, 400);
            } else if (evt.animationName == 'darken-right') {
               clearTimeout(timeout);
               passed = (passed && aboutEqual(delaySeen, 800));
               test(function(){assert_true(passed);}, testName); done();
               log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
            }
         });
         var delayStart = Date.now();
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should turn black, first the left half and then a
   little later the right half.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
